requirements.txt
PyQt5>=5.15.0
psutil>=5.9.0
speedtest-cli>=2.1.0
